// app/models/image-scrollable-visualization-data-model.ts
/**
 * @TO DO  
 *
 * Copyright (c) 2016, acme and/or its affiliates. All rights reserved.
 * 
 * @author Shafiq Abedin (sabedin@us.acme.com)
 * @version 1.0
 * @since 2016-9-1
 */

export class ImageScrollableVisualizationDataModel {
  id: number = 0;
  title: string = "";
  imageUrls: string[] = [];
  width: number = 100;
  height: number = 100;
  canvasWidth: number;
  canvasHeight: number;
  canDrawRectangle: boolean = false;

}